refactor: Rework broker clients to be more lightweight and only connect to exchanges and queues instead of creating them
